+++ src/main/java/io/electrum/moneytransfer/model/MoneyTransferAuthRequest.java
@@ -8,6 +8,7 @@ import javax.validation.constraints.NotNull;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import io.electrum.vas.Utils;
+import io.electrum.vas.model.Amounts;
 import io.electrum.vas.model.EncryptedPin;
 import io.electrum.vas.model.LedgerAmount;
 import io.electrum.vas.model.Originator;
@@ -41,18 +42,22 @@ public class MoneyTransferAuthRequest extends Transaction {
 
    private Relationship recipientRelationship = null;
 
+   private Amounts amounts = null;
+
    public MoneyTransferAuthRequest amount(LedgerAmount amount) {
       this.amount = amount;
       return this;
    }
 
    /**
-    * The amount to be transferred.
+    * The amount to be transferred. This field may be deprecated in a future version of the API. We encourage you to
+    * please also populate the 'amounts.requestAmount' field with this information.
     *
     * @return amount
     **/
    @JsonProperty("amount")
-   @ApiModelProperty(required = true, value = "The amount to be transferred.")
+   @ApiModelProperty(required = true, value = "The amount to be transferred. This field may be deprecated in a future "
+         + "version of the API. We encourage you to please also populate the 'amounts.requestAmount' field with this information.")
    @Valid
    @NotNull
    public LedgerAmount getAmount() {
@@ -251,6 +256,31 @@ public class MoneyTransferAuthRequest extends Transaction {
       this.recipientRelationship = relationship;
    }
 
+   public MoneyTransferAuthRequest amounts(Amounts amounts) {
+      this.amounts = amounts;
+      return this;
+   }
+
+   /**
+    * Amounts which make up the transaction. The existing 'amount' field currently takes precedence over this 'amounts'
+    * field, however the use of this 'amounts' field is encouraged. The 'amount' field may be deprecated in a future
+    * version of this API.
+    *
+    * @return amounts
+    **/
+   @JsonProperty("amounts")
+   @ApiModelProperty(value = "Amounts which make up the transaction. The existing 'amount' field currently takes "
+         + "precedence over this 'amounts' field, however the use of this 'amounts' field is encouraged. The 'amount' "
+         + "field may be deprecated in a future version of this API.")
+   @Valid
+   public Amounts getAmounts() {
+      return amounts;
+   }
+
+   public void setAmounts(Amounts amounts) {
+      this.amounts = amounts;
+   }
+
    @Override
    public boolean equals(Object o) {
       if (this == o)
@@ -265,7 +295,8 @@ public class MoneyTransferAuthRequest extends Transaction {
             && Objects.equals(customerProfileId, that.customerProfileId)
             && Objects.equals(newCustomer, that.newCustomer) && Objects.equals(fee, that.fee)
             && Objects.equals(quoteId, that.quoteId) && Objects.equals(purposeOfRemittance, that.purposeOfRemittance)
-            && Objects.equals(recipientRelationship, that.recipientRelationship);
+            && Objects.equals(recipientRelationship, that.recipientRelationship)
+            && Objects.equals(amounts, that.amounts);
    }
 
    @Override
@@ -281,7 +312,8 @@ public class MoneyTransferAuthRequest extends Transaction {
             fee,
             quoteId,
             purposeOfRemittance,
-            recipientRelationship);
+            recipientRelationship,
+            amounts);
    }
 
    @Override
@@ -307,6 +339,7 @@ public class MoneyTransferAuthRequest extends Transaction {
       sb.append("    quoteId: ").append(Utils.toIndentedString(quoteId)).append("\n");
       sb.append("    purposeOfRemittance: ").append(Utils.toIndentedString(purposeOfRemittance)).append("\n");
       sb.append("    recipientRelationship: ").append(Utils.toIndentedString(recipientRelationship)).append("\n");
+      sb.append("    amounts: ").append(Utils.toIndentedString(amounts)).append("\n");
       sb.append("}");
       return sb.toString();
    }