+++ src/main/java/io/electrum/moneytransfer/model/MoneyTransferRedeemResponse.java
@@ -8,6 +8,7 @@ import javax.validation.constraints.NotNull;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import io.electrum.vas.Utils;
+import io.electrum.vas.model.Amounts;
 import io.electrum.vas.model.LedgerAmount;
 import io.electrum.vas.model.Transaction;
 import io.swagger.annotations.ApiModel;
@@ -25,13 +26,17 @@ public class MoneyTransferRedeemResponse extends Transaction {
 
    private String customerProfileId = null;
 
+   private Amounts amounts = null;
+
    /**
-    * Get amount
+    * The amount to be transferred. This field may be deprecated in a future version of the API. We encourage you to
+    * please also populate the 'amounts.requestAmount' field with this information.
     * 
     * @return amount
     **/
    @JsonProperty("amount")
-   @ApiModelProperty(required = true, value = "")
+   @ApiModelProperty(required = true, value = "The amount to be transferred. This field may be deprecated in a future version of the API. "
+         + "We encourage you to please also populate the 'amounts.requestAmount' field with this information.")
    @Valid
    @NotNull
    public LedgerAmount getAmount() {
@@ -88,6 +93,31 @@ public class MoneyTransferRedeemResponse extends Transaction {
       this.customerProfileId = customerProfileId;
    }
 
+   public MoneyTransferRedeemResponse amounts(Amounts amounts) {
+      this.amounts = amounts;
+      return this;
+   }
+
+   /**
+    * Amounts which make up the transaction. The existing 'amount' field currently takes precedence over this 'amounts'
+    * field, however the use of this 'amounts' field is encouraged. The 'amount' field may be deprecated in a future
+    * version of this API.
+    *
+    * @return amounts
+    **/
+   @JsonProperty("amounts")
+   @ApiModelProperty(value = "Amounts which make up the transaction. The existing 'amount' field currently takes "
+         + "precedence over this 'amounts' field, however the use of this 'amounts' field is encouraged. The 'amount' "
+         + "field may be deprecated in a future version of this API.")
+   @Valid
+   public Amounts getAmounts() {
+      return amounts;
+   }
+
+   public void setAmounts(Amounts amounts) {
+      this.amounts = amounts;
+   }
+
    @Override
    public boolean equals(Object o) {
       if (this == o)
@@ -98,12 +128,12 @@ public class MoneyTransferRedeemResponse extends Transaction {
          return false;
       final MoneyTransferRedeemResponse that = (MoneyTransferRedeemResponse) o;
       return Objects.equals(amount, that.amount) && Objects.equals(orderId, that.orderId)
-            && Objects.equals(customerProfileId, that.customerProfileId);
+            && Objects.equals(customerProfileId, that.customerProfileId) && Objects.equals(amounts, that.amounts);
    }
 
    @Override
    public int hashCode() {
-      return Objects.hash(super.hashCode(), amount, orderId, customerProfileId);
+      return Objects.hash(super.hashCode(), amount, orderId, customerProfileId, amounts);
    }
 
    @Override
@@ -123,6 +153,7 @@ public class MoneyTransferRedeemResponse extends Transaction {
       sb.append("    slipData: ").append(Utils.toIndentedString(slipData)).append("\n");
       sb.append("    orderId: ").append(Utils.toIndentedString(orderId)).append("\n");
       sb.append("    customerProfileId: ").append(Utils.toIndentedString(customerProfileId)).append("\n");
+      sb.append("    amounts: ").append(Utils.toIndentedString(amounts)).append("\n");
       sb.append("}");
       return sb.toString();
    }